improve perforce support via awk processing
This change improves the perforce support by moving out the "p4 opened"
processing into it's own bash file that now has to be sourced at the top-level
bash-it.sh. Since the processing in simple the newly added: _p4-opened-counts
function returns a number of things that are not currently used, but since I had
awk open and doing the processing, I've chosen to include them in the output
anyway.

Testing:
 - Tested with the powerline-multiline theme
 - Ran:
     ❯ shellcheck themes/p4helpers.theme.bash
   and fixed all the errors